+++ lib/riif/dsl/klass.rb
@@ -1,0 +1,9 @@
+module Riif::DSL
+  class Klass < Base
+    HEADER_COLUMNS = [
+      :name
+    ]
+    START_COLUMN = 'CLASS'
+    END_COLUMN = ''
+  end
+end