+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.1 (unreleased)
 
+  * Implement CLASS
   * Implement ACCNT
   * Implement EMP
   * Implement CUST